+++ examples/sources/DocumentSource.js
@@ -54,10 +54,10 @@ class DocumentSource extends Component {
     }
 
     onRequestClose(e) {
-        const { onClose } = this.props;
+        const { onUpdate } = this.props;
         e.preventDefault();
 
-        onClose();
+        onUpdate();
     }
 
     onAfterOpen() {
@@ -107,7 +107,6 @@ DocumentSource.propTypes = {
     options: PropTypes.object.isRequired,
     entity: PropTypes.object,
     onUpdate: PropTypes.func.isRequired,
-    onClose: PropTypes.func.isRequired,
 };
 
 DocumentSource.defaultProps = {